requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.1.2
django-cleanup==5.1.0
idna==2.10
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.2.0
pycodestyle==2.6.0
pylint==2.6.0
PySocks==1.7.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
tweepy==3.9.0
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
psycopg2==2.7.7
